Make GreeterApplication a Spring Cloud Config Client
When GreeterApplication is started it will communicate with the Cloud
Config Server specified by `spring.cloud.config.uri` (Here it is
`http://127.0.0.1:8888`, the default is `http://localhost:8888`). The
values from this server take precedence.

The Config Server checks out a Git repository specified by
`spring.cloud.config.server.git.uri` and treats the files in it sort of
like they are in the Config Client's `config/` folder.

At the time of this commit, the `learn-spring-cloud-config-repo` repo
has only one file, `application.properties`. This means that every
application, regardless of its name, will get properties from this file.
(This initiall confused me because the pattern for doing an HTTP GET was
`/{application}/{profile}`. So regardless of what name you put values
from this file will be returned.) In the `application.properties` file
is `greetings.userName=Joey`, and indeed when the Config Server is
running and you run the GreeterApplication and go to
`http://localhost:8080` in your browser, you'll see "Hello, Joey!" as
opposed to "Hello, Bill!" (in the Greeting Application's
`application.properties`, `greetings.userName` is `Bill`).